# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/get_primary_service/{sortOrder}": {
            "get": {
                "tags": [
                    "Primary Services"
                ],
                "summary": "Primary service list",
                "operationId": "cc750a536a2581750a05aca3ffe0d9a6",
                "parameters": [
                    {
                        "name": "sortOrder",
                        "in": "query",
                        "description": "Primary service id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Primary Service List"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                }
            }
        },
        "/api/auth/store_primary_service": {
            "post": {
                "tags": [
                    "Primary Services"
                ],
                "summary": "Store primary service",
                "operationId": "5929b786daed36cfdf4254b048d89b42",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Primary service name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "business_key",
                        "in": "query",
                        "description": "Primary service business_key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "urlIcon": {
                                        "description": "Image file",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Primary Service Created"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/auth/update_primary_service/{id}": {
            "post": {
                "tags": [
                    "Primary Services"
                ],
                "summary": "Update Primary service",
                "operationId": "00b7214ccd5dd05213d3011dea28825b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Primary service id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Primary service name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "business_key",
                        "in": "query",
                        "description": "Primary service business_key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "urlIcon": {
                                        "description": "Image file",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Primary Service Updated"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/auth/delete_primary_service/{id}": {
            "get": {
                "tags": [
                    "Primary Services"
                ],
                "summary": "Delete primary service",
                "operationId": "6d30ac33dde032c1227e63606da489d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Primary service id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Primary Service"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}